War logic implemented. Coalitions and armies can fight now!

+++ src/common/GlobalConstants.java
@@ -3,7 +3,7 @@ package common;
 public final class GlobalConstants {
     public final static double ARMY_INITIAL_MORALE = 100;
 
-    public final static int SOLDIER_INITIAL_HEALTH = 100;
+    public final static double SOLDIER_INITIAL_HEALTH = 100;
 
 
 }
